Refactor: moved block filters to a separate module
